@@ -59,7 +59,7 @@ func populateMessage(message models.Message, msgType models.MessageType, channel
 // send - handles the sending logic of a message going to Discord.
 func send(dg *discordgo.Session, message models.Message, bot *models.Bot) {
 	if message.DirectMessageOnly {
-		err := handleDirectMessage(dg, message, bot)
+		err := handleDirectMessage(dg, message)
 		if err != nil {
 			log.Error().Msgf("problem sending message: %v", err)
 		}
@@ -72,7 +72,7 @@ func send(dg *discordgo.Session, message models.Message, bot *models.Bot) {
 }
 
 // handleDirectMessage - handle sending logic for direct messages.
-func handleDirectMessage(dg *discordgo.Session, message models.Message, bot *models.Bot) error {
+func handleDirectMessage(dg *discordgo.Session, message models.Message) error {
 	// Is output to rooms set?
 	if len(message.OutputToRooms) > 0 {
 		log.Warn().Msg("you have specified 'direct_message_only' as 'true' and provided 'output_to_rooms' -" +
